Show routePaths once when route is opened
