+++ src/stores/routeListStore.ts
@@ -41,22 +41,45 @@ export class RouteListStore {
     };
 
     @action
-    public toggleRoutePathVisibility = async (internalId: string) => {
+    public setRoutePathVisibility = async (isVisible: boolean, internalId: string) => {
         const currentRoutePath = this.getRoutePath(internalId);
-        if (currentRoutePath) {
-            currentRoutePath.visible = !currentRoutePath.visible;
-            currentRoutePath.color = currentRoutePath.visible
-                ? this.colorScale.reserveColor()
-                : this.colorScale.releaseColor(currentRoutePath.color!);
-            if (currentRoutePath.visible && currentRoutePath.routePathLinks.length === 0) {
-                const newRoutePath = await RoutePathService.fetchRoutePath(
-                    currentRoutePath.routeId,
-                    currentRoutePath.startTime,
-                    currentRoutePath.direction
-                );
-                this.updateRoutePathLinks(newRoutePath, internalId);
-            }
+        if (!currentRoutePath) return;
+        if (isVisible === currentRoutePath.visible) return;
+
+        currentRoutePath.visible = isVisible;
+        currentRoutePath.color = currentRoutePath.visible
+            ? this.colorScale.reserveColor()
+            : this.colorScale.releaseColor(currentRoutePath.color!);
+        if (currentRoutePath.visible && currentRoutePath.routePathLinks.length === 0) {
+            const newRoutePath = await RoutePathService.fetchRoutePath(
+                currentRoutePath.routeId,
+                currentRoutePath.startTime,
+                currentRoutePath.direction
+            );
+            this.updateRoutePathLinks(newRoutePath, internalId);
         }
+    };
+
+    @action
+    public toggleRoutePathVisibility = async (internalId: string) => {
+        console.log('toggleRoutePath visibility ', internalId);
+        const currentRoutePath = this.getRoutePath(internalId);
+        if (!currentRoutePath) return;
+        this.setRoutePathVisibility(!currentRoutePath.visible, internalId);
+        // if (currentRoutePath) {
+        //     currentRoutePath.visible = !currentRoutePath.visible;
+        //     currentRoutePath.color = currentRoutePath.visible
+        //         ? this.colorScale.reserveColor()
+        //         : this.colorScale.releaseColor(currentRoutePath.color!);
+        //     if (currentRoutePath.visible && currentRoutePath.routePathLinks.length === 0) {
+        //         const newRoutePath = await RoutePathService.fetchRoutePath(
+        //             currentRoutePath.routeId,
+        //             currentRoutePath.startTime,
+        //             currentRoutePath.direction
+        //         );
+        //         this.updateRoutePathLinks(newRoutePath, internalId);
+        //     }
+        // }
     };
 
     @action
